refactor(arch): move path.py to bloqade.lanes.arch (stage 6 of #569, non-breaking shim) (#576)
* refactor(arch): move ArchSpec + ArchSpecGeometry to bloqade.lanes.arch (stage 5 of #569)

Fifth stage of the multi-stage package merge tracked in #569. Moves
the canonical home of ArchSpec and ArchSpecGeometry / BusDescriptor
from bloqade.lanes.layout.{arch,arch_geometry} to
bloqade.lanes.arch.{spec,geometry} — runtime arch surface lives in
its dedicated package alongside the builders.

Old paths bloqade.lanes.layout.arch and bloqade.lanes.layout.arch_geometry
remain as thin re-export shims, so every existing caller across
bloqade.lanes, bloqade.gemini, demos, and tests compiles unchanged.
The shims are removed in the final cleanup stage (Stage 9 of the
plan) once all in-flight branches have rebased onto the canonical
paths.

Side fix: spec.py's relative `from .word import Word` would resolve
to bloqade.lanes.arch.word (which does not exist) post-move. Switched
to an absolute import `from bloqade.lanes.layout.word import Word`,
which works on main today and continues to work via Stage 2's shim
once that ships.

Refs #569.


* refactor(arch): move path.py to bloqade.lanes.arch (stage 6 of #569)

Sixth stage of the multi-stage package merge tracked in #569. Moves
the canonical home of PathFinder from bloqade.lanes.layout.path to
bloqade.lanes.arch.path — path-finding builds on top of the runtime
arch surface, so it lives alongside it.

Old path bloqade.lanes.layout.path remains as a thin re-export shim.
The shim is removed in the final cleanup stage.

Side fix: path.py's relative imports (.arch / .encoding / .move_metric)
would resolve to non-existent paths under bloqade.lanes.arch (until
later stages or shims land). Switched to absolute imports rooted at
bloqade.lanes.layout.* — these work on main today and continue to
work via the layout/ shims once each stage ships.

Refs #569.


* fix(arch): break circular import in stage 5 shims

bloqade.lanes.arch.arch_builder and .builder both imported
``ArchSpec`` from ``bloqade.lanes.layout.arch`` (the new shim). When
something does ``import bloqade.lanes.layout``, the shim triggers
``bloqade.lanes.arch.spec``, whose package-level ``__init__`` eagerly
loads arch_builder/builder, which in turn re-enter the partially-
loaded ``layout.arch`` shim — ImportError at the original entry.

Switch those modules (and the gemini sub-specs / validation that
referenced the same path) to import ``ArchSpec`` from
``bloqade.lanes.arch.spec`` directly. Six call sites updated:

- arch/arch_builder.py
- arch/builder.py
- arch/geometry.py (TYPE_CHECKING)
- arch/gemini/physical/spec.py
- arch/gemini/logical/spec.py
- arch/gemini/logical/validation.py

Also tightened the two shim docstrings (layout/arch.py and
layout/arch_geometry.py) to reference issue #569 instead of the
gitignored plan path that wasn't visible to readers of the
installed package.

Refs #569.


---------
