+++ python/bloqade/lanes/arch/__init__.py
@@ -2,6 +2,7 @@ from bloqade.lanes.arch.arch_builder import ArchBuilder, ZoneBuilder
 from bloqade.lanes.arch.builder import ArchResult, build_arch
 from bloqade.lanes.arch.geometry import ArchSpecGeometry, BusDescriptor
 from bloqade.lanes.arch.metrics import MoveMetricCalculator
+from bloqade.lanes.arch.path import PathFinder
 from bloqade.lanes.arch.spec import ArchSpec
 from bloqade.lanes.arch.topology import (
     AllToAllSiteTopology,
@@ -32,6 +33,7 @@ __all__ = [
     "InterZoneTopology",
     "MatchingTopology",
     "MoveMetricCalculator",
+    "PathFinder",
     "SiteTopology",
     "TransversalSiteTopology",
     "WordGrid",
